requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
tiktoken>=0.5.0
requests>=2.28.0
numpy>=1.20.0
tqdm>=4.65.0
python-dotenv>=1.0.0
openai>=0.27.0,<0.28.0